# api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Lucid API Reference",
    "contact": {
      "name": "Roost team",
      "url": "https://github.com/roostmoe/lucid",
      "email": "hello@roost.moe"
    },
    "license": {
      "name": "Apache 2.0 License",
      "url": "https://github.com/roostmoe/lucid/blob/main/LICENSE",
      "identifier": "Apache-2.0"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/cas": {
      "get": {
        "tags": [
          "cas"
        ],
        "operationId": "list_cas",
        "responses": {
          "200": {
            "description": "List of certificate authorities",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedList_Ca"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/schemas/ApiErrorResponse",
            "summary": "Client or validation error"
          },
          "401": {
            "$ref": "#/components/schemas/ApiErrorResponse",
            "summary": "Unauthorized"
          },
          "403": {
            "$ref": "#/components/schemas/ApiErrorResponse",
            "summary": "Forbidden"
          },
          "404": {
            "$ref": "#/components/schemas/ApiErrorResponse",
            "summary": "Not found"
          },
          "500": {
            "$ref": "#/components/schemas/ApiErrorResponse",
            "summary": "Internal server error"
          }
        }
      },
      "post": {
        "tags": [
          "cas"
        ],
        "summary": "Generate a new self-signed Ed25519 CA certificate and store it.",
        "operationId": "generate_ca",
        "responses": {
          "201": {
            "description": "Certificate authority generated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Ca"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/schemas/ApiErrorResponse",
            "summary": "Client or validation error"
          },
          "401": {
            "$ref": "#/components/schemas/ApiErrorResponse",
            "summary": "Unauthorized"
          },
          "403": {
            "$ref": "#/components/schemas/ApiErrorResponse",
            "summary": "Forbidden"
          },
          "404": {
            "$ref": "#/components/schemas/ApiErrorResponse",
            "summary": "Not found"
          },
          "500": {
            "$ref": "#/components/schemas/ApiErrorResponse",
            "summary": "Internal server error"
          }
        }
      }
    },
    "/api/v1/cas/import": {
      "post": {
        "tags": [
          "cas"
        ],
        "summary": "Import an existing CA certificate and private key.",
        "operationId": "import_ca",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImportCaRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Certificate authority imported",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Ca"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/schemas/ApiErrorResponse",
            "summary": "Client or validation error"
          },
          "401": {
            "$ref": "#/components/schemas/ApiErrorResponse",
            "summary": "Unauthorized"
          },
          "403": {
            "$ref": "#/components/schemas/ApiErrorResponse",
            "summary": "Forbidden"
          },
          "404": {
            "$ref": "#/components/schemas/ApiErrorResponse",
            "summary": "Not found"
          },
          "500": {
            "$ref": "#/components/schemas/ApiErrorResponse",
            "summary": "Internal server error"
          }
        }
      }
    },
    "/api/v1/cas/{id}": {
      "get": {
        "tags": [
          "cas"
        ],
        "operationId": "get_ca",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Certificate authority details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Ca"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/schemas/ApiErrorResponse",
            "summary": "Client or validation error"
          },
          "401": {
            "$ref": "#/components/schemas/ApiErrorResponse",
            "summary": "Unauthorized"
          },
          "403": {
            "$ref": "#/components/schemas/ApiErrorResponse",
            "summary": "Forbidden"
          },
          "404": {
            "$ref": "#/components/schemas/ApiErrorResponse",
            "summary": "Not found"
          },
          "500": {
            "$ref": "#/components/schemas/ApiErrorResponse",
            "summary": "Internal server error"
          }
        }
      },
      "delete": {
        "tags": [
          "cas"
        ],
        "operationId": "delete_ca",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Certificate authority deleted"
          },
          "400": {
            "$ref": "#/components/schemas/ApiErrorResponse",
            "summary": "Client or validation error"
          },
          "401": {
            "$ref": "#/components/schemas/ApiErrorResponse",
            "summary": "Unauthorized"
          },
          "403": {
            "$ref": "#/components/schemas/ApiErrorResponse",
            "summary": "Forbidden"
          },
          "404": {
            "$ref": "#/components/schemas/ApiErrorResponse",
            "summary": "Not found"
          },
          "500": {
            "$ref": "#/components/schemas/ApiErrorResponse",
            "summary": "Internal server error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiErrorResponse": {
        "type": "object",
        "description": "An error response for an API endpoint. This is used to return errors to the\nclient in a consistent format.",
        "required": [
          "message"
        ],
        "properties": {
          "code": {
            "type": [
              "string",
              "null"
            ],
            "description": "An optional error code that can be used to identify the type of error\nthat occurred."
          },
          "details": {
            "type": [
              "string",
              "null"
            ]
          },
          "message": {
            "type": "string",
            "description": "A human-readable message describing the error that occurred."
          }
        }
      },
...
      "Ca": {
        "type": "object",
        "description": "A certificate authority managed by Lucid.",
        "required": [
          "id",
          "cert_pem",
          "fingerprint",
          "created_at"
        ],
        "properties": {
          "cert_pem": {
            "type": "string",
            "description": "CA certificate in PEM format"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "When this CA was created"
          },
          "fingerprint": {
            "type": "string",
            "description": "SHA-256 fingerprint of the certificate (format: `sha256:<hex>`)"
          },
          "id": {
            "type": "string",
            "description": "Internal database ID"
          }
        }
      },
...
      "ImportCaRequest": {
        "type": "object",
        "description": "Request body for importing an existing certificate authority.",
        "required": [
          "cert_pem",
          "private_key_pem"
        ],
        "properties": {
          "cert_pem": {
            "type": "string",
            "description": "PEM-encoded CA certificate."
          },
          "private_key_pem": {
            "type": "string",
            "description": "PEM-encoded private key for the CA certificate (will be encrypted at\nrest using the server's `LUCID_CA_ENCRYPTION_KEY`)."
          }
        }
      },
...
      "PaginatedList_Ca": {
        "type": "object",
        "description": "Parameters for paginating through a list of records. This is used by the\nvarious list endpoints to allow clients to paginate through large sets of\nrecords.",
        "required": [
          "items"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "type": "object",
              "description": "A certificate authority managed by Lucid.",
              "required": [
                "id",
                "cert_pem",
                "fingerprint",
                "created_at"
              ],
              "properties": {
                "cert_pem": {
                  "type": "string",
                  "description": "CA certificate in PEM format"
                },
                "created_at": {
                  "type": "string",
                  "format": "date-time",
                  "description": "When this CA was created"
                },
                "fingerprint": {
                  "type": "string",
                  "description": "SHA-256 fingerprint of the certificate (format: `sha256:<hex>`)"
                },
                "id": {
                  "type": "string",
                  "description": "Internal database ID"
                }
              }
            }
          },
          "limit": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "The maximum number of results to return.",
            "minimum": 0
          },
          "next_token": {
            "type": [
              "string",
              "null"
            ],
            "description": "The next page token, if any. This is acquired by requesting a paginated\nset of records and looking at the `next_token` or `prev_token` field."
          }
        }
      },
...
    }
  }
}